# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EDU SMART OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: circle_id
          in: query
          schema: { type: string }
        - name: level
          in: query
          schema: { type: string }
        - name: payment_status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /teachers:
    get:
      operationId: listTeachers
      tags: [teachers]
      summary: List all teachers
      responses:
        "200":
          description: List of teachers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
    post:
      operationId: createTeacher
      tags: [teachers]
      summary: Create a teacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
  /teachers/{id}:
    get:
      operationId: getTeacher
      tags: [teachers]
      summary: Get a teacher by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
    patch:
      operationId: updateTeacher
      tags: [teachers]
      summary: Update a teacher
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherUpdate"
      responses:
        "200":
          description: Updated teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
    delete:
      operationId: deleteTeacher
      tags: [teachers]
      summary: Delete a teacher
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: type
          in: query
          schema: { type: string }
        - name: unread
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, full_name, grade, enrollment_date]
      properties:
        id: { type: string }
        full_name: { type: string }
        photo_url: { type: ["string", "null"] }
        age: { type: ["integer", "null"] }
        birth_date: { type: ["string", "null"] }
        grade: { type: string }
        address: { type: ["string", "null"] }
        governorate: { type: ["string", "null"] }
        guardian_phone: { type: string }
        secondary_phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        teacher_id: { type: ["string", "null"] }
        teacher_name: { type: ["string", "null"] }
        circle_id: { type: ["string", "null"] }
        circle_name: { type: ["string", "null"] }
        payment_status: { type: string }
        payment_amount: { type: ["number", "null"] }
        is_exempt: { type: boolean }
        current_memorization: { type: ["string", "null"] }
        current_revision: { type: ["string", "null"] }
        level: { type: ["string", "null"] }
        rating: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
        enrollment_date: { type: string }
        points: { type: integer }
        created_at: { type: string }
        updated_at: { type: string }
    StudentInput:
      type: object
      required: [full_name, grade, guardian_phone]
      properties:
        full_name: { type: string }
        photo_url: { type: string }
        age: { type: integer }
        birth_date: { type: string }
        grade: { type: string }
        address: { type: string }
        governorate: { type: string }
        guardian_phone: { type: string }
        secondary_phone: { type: string }
        email: { type: string }
        teacher_id: { type: string }
        circle_id: { type: string }
        payment_status: { type: string }
        payment_amount: { type: number }
        is_exempt: { type: boolean }
        current_memorization: { type: string }
        current_revision: { type: string }
        level: { type: string }
        rating: { type: integer }
        notes: { type: string }
        enrollment_date: { type: string }
    StudentUpdate:
      type: object
      properties:
        full_name: { type: string }
        photo_url: { type: string }
        age: { type: integer }
        birth_date: { type: string }
        grade: { type: string }
        address: { type: string }
        governorate: { type: string }
        guardian_phone: { type: string }
        secondary_phone: { type: string }
        email: { type: string }
        teacher_id: { type: string }
        circle_id: { type: string }
        payment_status: { type: string }
        payment_amount: { type: number }
        is_exempt: { type: boolean }
        current_memorization: { type: string }
        current_revision: { type: string }
        level: { type: string }
        rating: { type: integer }
        notes: { type: string }
    Teacher:
      type: object
      required: [id, full_name, phone]
      properties:
        id: { type: string }
        full_name: { type: string }
        photo_url: { type: ["string", "null"] }
        age: { type: ["integer", "null"] }
        phone: { type: string }
        secondary_phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        salary: { type: ["number", "null"] }
        hire_date: { type: ["string", "null"] }
        experience: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        student_count: { type: integer }
        circle_count: { type: integer }
        created_at: { type: string }
        updated_at: { type: string }
    TeacherInput:
      type: object
      required: [full_name, phone]
      properties:
        full_name: { type: string }
        photo_url: { type: string }
        age: { type: integer }
        phone: { type: string }
        secondary_phone: { type: string }
        email: { type: string }
        address: { type: string }
        salary: { type: number }
        hire_date: { type: string }
        experience: { type: string }
        notes: { type: string }
    TeacherUpdate:
      type: object
      properties:
        full_name: { type: string }
        photo_url: { type: string }
        age: { type: integer }
        phone: { type: string }
        secondary_phone: { type: string }
        email: { type: string }
        address: { type: string }
        salary: { type: number }
        hire_date: { type: string }
        experience: { type: string }
        notes: { type: string }
...
    Notification:
      type: object
      required: [id, type, title, message, is_read, created_at]
      properties:
        id: { type: string }
        type: { type: string }
        title: { type: string }
        message: { type: string }
        student_id: { type: ["string", "null"] }
        student_name: { type: ["string", "null"] }
        is_read: { type: boolean }
        priority: { type: ["string", "null"] }
        created_at: { type: string }
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        id: { type: string }